requirements.txt
uwsgi
python-dotenv
flask
pymongo
certifi
PyJWT